Update FPPDiscovery.cpp
Fix for defect where an empty (null) hostname in the WiFi class causes an exception. This will happen "out of the box" when the config file hasn't been setup with a hostname and there isn't one specified in the ESPixelStick.ino setup code.

We could default to something in the INO but this is a good change anyway since it will protect folks in case they accidently remove the default from the caller.