Fixed issue in the status response that returned a string when an enum was expected. This caused odd mode reporting including "unknown" mode.
Fixed detection of FPP pings that was broken when the ping requests were seperated due to an endless loop of pings. FPP  sends a specific ping subtype. The remotes need to respond with an alternative sub type that will not trigger the other remote devices to generate pings. The correct fix should have been to send pings from the ESP using the remote device ping sub type.

Fixed remote vs bridge reporting to be aware of FPP settings in the config as well as file selection in the config. Local autoplay still trumps FPP master.

Added ping generation when the WiFi interface declares itself operational. This triggers the Master to talk to us. This fixes a race condition with the initial ping trying to go out before the WiFi is up.